+++ src/Sequence/Defer.php
@@ -637,6 +637,31 @@ final class Defer implements Implementation
     }
 
     /**
+     * @template I
+     *
+     * @param I $carry
+     * @param callable(I, T, Sink\Continuation<I>): Sink\Continuation<I> $reducer
+     *
+     * @return I
+     */
+    public function sink($carry, callable $reducer): mixed
+    {
+        $continuation = Sink\Continuation::of($carry);
+
+        foreach ($this->values as $value) {
+            /** @psalm-suppress ImpureFunctionCall */
+            $continuation = $reducer($carry, $value, $continuation);
+            $carry = $continuation->unwrap();
+
+            if (!$continuation->shouldContinue()) {
+                return $continuation->unwrap();
+            }
+        }
+
+        return $continuation->unwrap();
+    }
+
+    /**
      * @return Implementation<T>
      */
     public function clear(): Implementation