+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 5.11.0 - 2024-12-01
+
+### Added
+
+- `Innmind\Immutable\Sequence::sink()`
+- `Innmind\Immutable\Attempt`
+
+### Fixed
+
+- `Innmind\Immutable\Maybe::memoize()` and `Innmind\Immutable\Either::memoize()` was only unwrapping the first layer of the monad. It now recursively unwraps until all the deferred monads are memoized.
+
 ## 5.10.0 - 2024-11-09
 
 ### Added